+++ modules/data_dictionary_widget/src/Indexes/IndexValidation.php
@@ -42,23 +42,23 @@ class IndexValidation {
    * @param string $field_label
    *   The label for the field being edited.
    */
-  public static function indexFieldVal(FormStateInterface $form_state, string $field_key, string $field_label) {
-    $trigger = $form_state->getTriggeringElement();
-    $op = $trigger['#op'];
-    // We split the key to get the index field location.
-    $op_index = explode('_', $op);
-
-    // Perform validation for update operation.
-    if (str_contains($op, 'update')) {
-      $update_values = $form_state->getUserInput();
-      $value = $update_values["field_json_metadata"][0]["dictionary_fields"]["data"][$op_index[1]]["field_collection"][$field_key];
-      if ($value === "") {
-        $field = "field_json_metadata][0][indexes][fields][edit_fields][$op_index[1]][$field_key";
-        $form_state->setErrorByName($field, t(':field_label field is required.', [
-          ':field_label' => $field_label,
-        ]));
-      }
-    }
-  }
+//  public static function indexFieldVal(FormStateInterface $form_state, string $field_key, string $field_label) {
+//    $trigger = $form_state->getTriggeringElement();
+//    $op = $trigger['#op'];
+//    // We split the key to get the index field location.
+//    $op_index = explode('_', $op);
+//
+//    // Perform validation for update operation.
+//    if (str_contains($op, 'update')) {
+//      $update_values = $form_state->getUserInput();
+//      $value = $update_values["field_json_metadata"][0]["dictionary_fields"]["data"][$op_index[1]]["field_collection"][$field_key];
+//      if ($value === "") {
+//        $field = "field_json_metadata][0][indexes][fields][edit_fields][$op_index[1]][$field_key";
+//        $form_state->setErrorByName($field, t(':field_label field is required.', [
+//          ':field_label' => $field_label,
+//        ]));
+//      }
+//    }
+//  }
 
 }
